Create getlivery.vbs add scripts/getlivery.vbs to a generate livery list ready for upload to an instance closes #8
Proper command line usage of this script is:
	getlivery [-beta] [filename]

where the -beta argument will parse the DCS World Openbeta main folder and user folder
and filename is the name of the output file.
If no path is specified, the file will be output to the DCS user folder (main or openbeta depending on -beta argument).
If no filename is specified, the output will be "DCS_Liveries.txt" in the DCS user folder.
If you want the filename created in the same folder as this script, use the following line:
	getlivery [-beta] .\filename

IMPORTANT: if you use the -beta argument, it must come before the filename.