When DCS: World map is not found, fall back to online version
I have put up a copy of the DCS: World map data at
http://dcs-mission-planner.combined-ops-group.com:8100/,
mostly to make the DCS: World map available to the online version of the
mission planner.

I may have to take it down again if it consumes too much traffic, but
with the small userbase we have now and taking browser caches into account
this should hold for a while.